tests(runloop) adjust plugin iterator invalidation tests for timer
The logic for invalidating / updating the plugins iterator has changed due to the introduction of timer-based updates - In particular, the `plugins_iterator:version` cache entry is `nil`  no more, it's always `init` or a uuid. When a node makes a change which requires updating the plugins iterator, this change gets "noticed" by the other nodes automatically, without having to wait for "activity".

Co-Authored-By: Aapo Talvensaari <bungle@konghq.com>